+++ src/QuickMarcEditor/QuickMarcEditorRows/FixedField/ComputerFileFixedField.js
@@ -91,5 +91,6 @@ ComputerFileFixedField.propTypes = {
 };
 
 ComputerFileFixedField.displayName = 'ComputerFileFixedField';
+ComputerFileFixedField.configFields = config.fields;
 
 export default ComputerFileFixedField;