+++ src/QuickMarcEditor/QuickMarcEditorRows/PhysDescriptionField/GlobePhysDescriptionField.js
@@ -48,4 +48,6 @@ GlobePhysDescriptionField.propTypes = {
   name: PropTypes.string.isRequired,
 };
 
+GlobePhysDescriptionField.configFields = configFields;
+
 export default GlobePhysDescriptionField;